+++ src/main/java/lotto/domain/Lottos.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 import static java.util.stream.Collectors.*;
